Port QueueSpec
